default-values-and-collections: Restructure `api` module
