+++ api/src/main/kotlin/ru/potatophobe/konfig/api/Konfig.kt
@@ -1,0 +1,13 @@
+package ru.potatophobe.konfig.api
+
+import kotlin.reflect.KClass
+
+/**
+ * Container for declared Konfigs
+ *
+ * @see KonfigClass
+ * */
+interface Konfig {
+    fun <T : Any> get(kClass: KClass<T>): T
+    fun <T : Any> getOrNull(kClass: KClass<T>): T?
+}
