+++ api/src/main/kotlin/ru/potatophobe/konfig/api/KonfigDsl.kt
@@ -1,0 +1,9 @@
+package ru.potatophobe.konfig.api
+
+/**
+ * Marks Konfig DSL components
+ * */
+@DslMarker
+@MustBeDocumented
+@Retention(AnnotationRetention.BINARY)
+annotation class KonfigDsl
