+++ api/src/main/kotlin/ru/potatophobe/konfig/api/KonfigScope.kt
@@ -1,0 +1,11 @@
+package ru.potatophobe.konfig.api
+
+/**
+ * Implicit receiver for all generated Konfig DSL root functions
+ *
+ * @see KonfigClass
+ * */
+@KonfigDsl
+interface KonfigScope {
+    fun add(konfig: Any)
+}
